[spring-web-mvc] add GET /cart/items

+++ src/main/java/edu/yandex/project/controller/dto/CartView.java
@@ -2,9 +2,9 @@ package edu.yandex.project.controller.dto;
 
 import java.util.List;
 
-public record CartView(List<ItemView> items, int totalPrice) {
+public record CartView(List<ItemView> items, long totalPrice) {
 
     public static CartView createStub() {
-        return new CartView(List.of(), 0);
+        return new CartView(List.of(), 0L);
     }
 }
